feat(ddsql): migrate execution to public v2 API
Route DDSQL table, time-series, and security findings analysis through the public tabular query and fetch endpoints. Validate time windows and row limits client-side, poll the v2 state field, and parse the single-object column-major response shape.\n\nAdd request, response, polling, OAuth, User-Agent, and HTTP error coverage for the migrated contract.

+++ src/commands/security.rs
@@ -138,7 +138,7 @@ pub async fn findings_analyze(
         query,
         from,
         to,
-        Some(limit as i32),
+        Some(limit),
         Some("security-findings-analyze"),
     )
     .await
